pyproject.toml
[project]
name = "gruppen-academy"
version = "0.1.0"
description = "Plataforma interna de aprendizagem corporativa da Gruppen"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "pydantic[email]>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "python-jose[cryptography]>=3.3,<4",
    "passlib[bcrypt]>=1.7,<2",
    "bcrypt>=4.0,<4.1",
    "python-multipart>=0.0.18",
    "openai>=1.60,<2",
    "httpx>=0.28,<1",
    "fpdf2>=2.8,<3",
    "pdfplumber>=0.11,<1",
    "pytesseract>=0.3.10,<1",
    "pdf2image>=1.16,<2",
    "Pillow>=10.0,<11",
    "qrcode>=8.0,<9",
    "pyzbar>=0.1.9,<1",
    "redis>=5.0,<6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
